refactor: provider 분리, components/NotificationProvider.tsx 분리
